move candidate component into consensus package
