+++ pkg/core/consensus/candidate/blockgenerator.go
