+++ pkg/core/consensus/candidate/factory.go
