+++ pkg/core/consensus/candidate/generation_test.go
@@ -3,8 +3,8 @@ package candidate_test
 import (
 	"testing"
 
-	"github.com/dusk-network/dusk-blockchain/pkg/core/chain/candidate"
 	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus"
+	"github.com/dusk-network/dusk-blockchain/pkg/core/consensus/candidate"
 	"github.com/dusk-network/dusk-blockchain/pkg/core/marshalling"
 	"github.com/dusk-network/dusk-blockchain/pkg/util/nativeutils/eventbus"
 	"github.com/dusk-network/dusk-blockchain/pkg/util/nativeutils/rpcbus"
