+++ pkg/core/consensus/candidate/genesis.go
