+++ pkg/core/consensus/candidate/genesis_test.go
