+++ pkg/core/consensus/candidate/readme.md
