+++ pkg/core/consensus/candidate/testutil.go
